CID 1509059 winbind: Fixing print statement for time_t
Fixing DEBUG output for time_t to uintmax_t


Autobuild-User(master): Martin Schwenke <martins@samba.org>
Autobuild-Date(master): Wed Aug 27 02:17:29 UTC 2025 on atb-devel-224

+++ source3/passdb/passdb.c
@@ -2226,8 +2226,8 @@ bool pdb_update_autolock_flag(struct samu *sampass, bool *updated)
 	}
 
 	LastBadPassword = pdb_get_bad_password_time(sampass);
-	DEBUG(7, ("pdb_update_autolock_flag: Account %s, LastBadPassword=%d, duration=%d, current time =%d.\n",
-		  pdb_get_username(sampass), (uint32_t)LastBadPassword, duration*60, (uint32_t)time(NULL)));
+	DEBUG(7, ("pdb_update_autolock_flag: Account %s, LastBadPassword=%ju, duration=%d, current time =%ju.\n",
+		  pdb_get_username(sampass), (uintmax_t)LastBadPassword, duration*60, (uintmax_t)time(NULL)));
 
 	if (LastBadPassword == (time_t)0) {
 		DEBUG(1,("pdb_update_autolock_flag: Account %s "
